change int to float for prices

+++ src/Film.java
@@ -1,7 +1,7 @@
 //Léo
 public class Film {
 
-    private int prix;
+    private float prix;
     private String titre;
     private GlobalVals.categories categorie;
     private GlobalVals.genres genre;
@@ -9,24 +9,24 @@ public class Film {
     private Film film;
 
     public Film() {
-        this.prix = 0;
+        this.prix = 0f;
         this.titre = "";
         this.categorie = GlobalVals.categories.Ancien;
         this.genre = GlobalVals.genres.Fantaisie;
     }
 
-    public Film(int prix, String titre, GlobalVals.categories categorie, GlobalVals.genres genre) {
+    public Film(float prix, String titre, GlobalVals.categories categorie, GlobalVals.genres genre) {
         this.prix = prix;
         this.titre = titre;
         this.categorie = categorie;
         this.genre = genre;
     }
 
-    public int getPrix() {
+    public float getPrix() {
         return prix;
     }
 
-    public void setPrix(int prix) {
+    public void setPrix(float prix) {
         this.prix = prix;
     }
